backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.1.6",
    "django-cors-headers>=4.3.1",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.3.0",
    "pillow>=10.1.0",
]
